requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
asgiref==3.7.2
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.11.2
black==22.1.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.0.4
colorama==0.4.4
comtypes==1.2.0
ConfigArgParse==1.7
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.2
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==0.8.2
filetype==1.2.0
Flask==2.2.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
frozenlist==1.3.3
gevent==24.2.1
geventhttpclient==2.2.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib-metadata==6.5.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
l18n==2021.3
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-google-genai==1.0.7
langchain-openai==0.1.12
langchain-text-splitters==0.2.2
langserve==0.2.2
langsmith==0.1.82
locust==2.27.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.3
matplotlib-inline==0.1.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==0.4.3
numpy==1.21.4
openai==1.35.7
openpyxl==3.1.2
orjson==3.10.5
packaging==24.1
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pdfrw==0.4
photoshop-python-api==0.21.7
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pygame==2.1.0
pygames==0.1
Pygments==2.18.0
pyparsing==3.0.7
PyPDF2==3.0.1
pyproject-toml==0.0.10
PySimpleGUI==4.56.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
secure-smtplib==0.1.1
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==2.0.31
sqlparse==0.4.4
stack-data==0.2.0
streamlit==1.36.0
sympy==1.11.1
telepath==0.3.1
tenacity==8.4.2
threadpoolctl==3.0.0
tiktoken==0.7.0
tmdbv3api==1.9.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==1.10.2
tornado==6.4.1
tqdm==4.65.0
traitlets==5.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vboxapi==1.0
wagtail==5.0.2
watchdog==4.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.3
Willow==1.5
yarl==1.8.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.3
